+++ src/pas/plugins/imio/upgrades.py
@@ -1,5 +1,7 @@
+from Acquisition import aq_base
 from BTrees.OOBTree import OOBTree
 from plone import api
+from Products.CMFCore.interfaces import IContentish
 from pas.plugins.imio.browser.view import AddAuthenticUsers
 from authomatic.core import User
 
@@ -10,6 +12,7 @@ logger = logging.getLogger(__file__)
 
 def set_new_userid(context=None):
     portal = api.portal.get()
+    catalog = api.portal.get_tool("portal_catalog")
     view = AddAuthenticUsers(portal, portal.REQUEST)
     users = view.get_authentic_users()
 
@@ -49,3 +52,57 @@ def set_new_userid(context=None):
         logger.info(
             "user updated, new id is:{}, new login is: {}".format(userid, userlogin)
         )
+
+    def convert_userid(login):
+        user = plugin._useridentities_by_login.get(login)
+        return user and user.userid or login
+
+    def convert_userids(logins):
+        result = []
+        for login in logins:
+            userid = convert_userid(login)
+            result.append(userid)
+        return result
+
+    def do_migrate_roles(obj, path):
+        obj_url = obj.absolute_url()
+        if not IContentish.providedBy(obj):
+            return
+
+        # migrate local roles
+        if getattr(aq_base(obj), "__ac_local_roles__", None) is not None:
+            localroles = obj.__ac_local_roles__
+            migrated = False
+            for login in localroles:
+                roles = localroles[login]
+                userid = convert_userid(login)
+                if userid == login:
+                    continue
+                obj.manage_delLocalRoles([login])
+                obj.manage_setLocalRoles(userid=userid, roles=roles)
+                migrated = True
+            if migrated:
+                logger.info(u"Migrated userids in local roles on {}".format(obj_url))
+
+        # migrate creators
+        creators = getattr(obj, 'listCreators', [])
+        if callable(creators):
+            creators = creators()
+        new_creators = tuple(convert_userids(creators))
+        if creators != new_creators:
+            obj.setCreators(new_creators)
+            obj.reindexObject(idxs=["Creator", "listCreators"])
+            logger.info(u"Migrated creator(s) on {}".format(obj_url))
+
+        # migrate contributors
+        contributors = getattr(obj, 'listContributors', [])
+        if callable(contributors):
+            contributors = contributors()
+        new_contributors = tuple(convert_userids(contributors))
+        if contributors != new_contributors:
+            obj.setContributors(new_contributors)
+            logger.info(u"Migrated contributors(s) on {}".format(obj_url))
+
+    portal.ZopeFindAndApply(portal, search_sub=True, apply_func=do_migrate_roles)
+    catalog.reindexIndex("allowedRolesAndUsers", None)
+    logger.info("Reindexed security")